Refactor: Extract network dependencies to separate Koin module
This commit extracts Retrofit and OkHttpClient configurations from `dataModule.kt` into a new `networkModule.kt`.

The `networkModule` now defines:
- An "authRetrofit" instance for the WhatsApp API, including API key and content type headers.
- A "locationRetrofit" instance for the countries API, including a content type header.
- Corresponding "authOkHttp" and "locationOkHttp" OkHttpClient instances.

The `appModule.kt` has been updated to include the new `networkModule`. This improves modularity and organization of dependency injection.
