+++ app/src/main/java/com/paris_2/san3a/di/appModule.kt
@@ -3,5 +3,13 @@ package com.paris_2.san3a.di
 import org.koin.dsl.module
 
 val appModule = module {
-    includes(navigationModule, dataModule,repositoryModule,useCaseModule,viewModelModule,uiModule)
+    includes(
+        navigationModule,
+        dataModule,
+        repositoryModule,
+        useCaseModule,
+        viewModelModule,
+        uiModule,
+        networkModule
+    )
 }
